CatCyberOverlord refactored into separate file
